==> python/2_var_creation.ipynb <==
# Import dependencies
import pandas as pd
import numpy as np

# Read in the Scripts CSV file
cleaned_df = pd.read_csv('../outputData/cleaned_scripts.csv')

# Drop the "Unnamed: 0" and "Unnamed: 0.1" columns
cleaned_df = cleaned_df.drop("Unnamed: 0", 1)
cleaned_df = cleaned_df.drop("Unnamed: 0.1", 1)
cleaned_df.head()

# Create variables for WordCount and LineCount
# LineCount is easy - just make it equal to 1, then when we group by episode, take the sum

# LineCount variable created
cleaned_df["LineCount"] = np.repeat(1, cleaned_df.shape[0])
cleaned_df.head()

#Create a function for counting the # of words per line of dialogue
def word_count(text):
    count = 0
    words = text.split()

    for word in words:
        count += 1
    
    return count

# WordCount -- create a loop through each row and each Dialogue...

# Create a list that will later be added to dataframe
WordCount = []

for i in range(cleaned_df.shape[0]):
    
    # Assign the current line of text to variable 'line'
    line = str(cleaned_df["Dialogue"][i]).lower()
    
    # Run the word_count function on line i
    try:
        count = word_count(line)
    except:
        count = 0
    
    # Append count to WordCount list
    WordCount.append(count)
    
    #########################
    # Print notifications to make sure the code is running
    #########################

    if i % 10000 == 0:
        perc_complete = round(i / cleaned_df.shape[0] * 100, 0)
        print("Percent Complete: " + str(perc_complete) + "%")
    
    if i == (cleaned_df.shape[0] - 1):
        print("-----------------------------------")
        print("Loop Complete!")

#Add WordCount as a column in dataframe
cleaned_df["WordCount"] = WordCount
cleaned_df.tail()

# Create a weighted sentiment score (Sentiment x WordCount)
cleaned_df["WtdPos"] = cleaned_df["PosSent"] * cleaned_df["WordCount"]
cleaned_df["WtdNeg"] = cleaned_df["CompSent"] * cleaned_df["WordCount"]
cleaned_df["WtdComp"] = cleaned_df["NegSent"] * cleaned_df["WordCount"]
cleaned_df.tail(10)

##### Create columns for Line Count, Word Count and Wtd Sentiment for each character / character group
# 1) Jerry
# 2) George
# 3) Kramer
# 4) Elaine
# 5) Secondary Characters
# 6) Other Characters

# Use loc to separate out each CHARACTER into separate dfs

#Add list for main characters
mainChars = ["JERRY", "GEORGE", "KRAMER", "ELAINE"]

# Create separate dataframes 
jerry_df = cleaned_df.loc[cleaned_df["Character"] == "JERRY", :]
george_df = cleaned_df.loc[cleaned_df["Character"] == "GEORGE", :]
kramer_df = cleaned_df.loc[cleaned_df["Character"] == "KRAMER", :]
elaine_df = cleaned_df.loc[cleaned_df["Character"] == "ELAINE", :]

mainChars_df = cleaned_df.loc[(cleaned_df["Character"] == "JERRY") | 
                             (cleaned_df["Character"] == "GEORGE") |
                             (cleaned_df["Character"] == "KRAMER") |
                             (cleaned_df["Character"] == "ELAINE"), :]

secondary_df = cleaned_df.loc[cleaned_df["Character"] == "SECONDARY", :]
other_df = cleaned_df.loc[cleaned_df["Character"] == "OTHER", :]

# Group dataframes (including the original cleaned_df) by Episode

# Grouping by Episode -- Overall (a.k.a. all characters)
overall_grp = cleaned_df.groupby(by = "SEID", as_index=False).sum()

# Drop the EpisodeNo, Season, and Sentiment columns since we don't need them anymore

# Re-create columns for EpisodeNo, Season and Sentiment
overall_grp["EpisodeNo"] = overall_grp["EpisodeNo"] / overall_grp["LineCount"]
overall_grp["Season"] = overall_grp["Season"] / overall_grp["LineCount"]
overall_grp["PosSent"] = overall_grp["WtdPos"] / overall_grp["WordCount"]
overall_grp["NegSent"] = overall_grp["WtdNeg"] / overall_grp["WordCount"]
overall_grp["CompSent"] = overall_grp["WtdComp"] / overall_grp["WordCount"]

# Drop column for WtdSentiments since we no longer need it
overall_grp = overall_grp.drop("WtdPos", 1)
overall_grp = overall_grp.drop("WtdNeg", 1)
overall_grp = overall_grp.drop("WtdComp", 1)

# Rename Sentiment, LineCount, and WordCount variables to have "Total" at the end of them
overall_grp = overall_grp.rename(index=str, columns = {"PosSent": "PosTotal",
                                                       "NegSent": "NegTotal",
                                                       "CompSent": "CompTotal",
                                                       "LineCount": "LineCountTotal",
                                                       "WordCount": "WordCountTotal"
                                                      })
overall_grp.tail()

# Grouping by Episode -- Jerry
jerry_grp = jerry_df.groupby(by = "SEID", as_index=False).sum()

# Drop the EpisodeNo, Season, and Sentiment columns since we don't need them anymore

# Re-create columns for EpisodeNo, Season and Sentiment
jerry_grp["EpisodeNo"] = jerry_grp["EpisodeNo"] / jerry_grp["LineCount"]
jerry_grp["Season"] = jerry_grp["Season"] / jerry_grp["LineCount"]
jerry_grp["PosSent"] = jerry_grp["WtdPos"] / jerry_grp["WordCount"]
jerry_grp["NegSent"] = jerry_grp["WtdNeg"] / jerry_grp["WordCount"]
jerry_grp["CompSent"] = jerry_grp["WtdComp"] / jerry_grp["WordCount"]

# Drop column for WtdSentiments since we no longer need it
jerry_grp = jerry_grp.drop("WtdPos", 1)
jerry_grp = jerry_grp.drop("WtdNeg", 1)
jerry_grp = jerry_grp.drop("WtdComp", 1)

# Rename Sentiment, LineCount, and WordCount variables to have "Total" at the end of them
jerry_grp = jerry_grp.rename(index=str, columns = {"PosSent": "PosJerry",
                                                       "NegSent": "NegJerry",
                                                       "CompSent": "CompJerry",
                                                       "LineCount": "LineCountJerry",
                                                       "WordCount": "WordCountJerry"
                                                      })
jerry_grp.tail()

# Grouping by Episode -- George
george_grp = george_df.groupby(by = "SEID", as_index=False).sum()

# Drop the EpisodeNo, Season, and Sentiment columns since we don't need them anymore

# Re-create columns for EpisodeNo, Season and Sentiment
george_grp["EpisodeNo"] = george_grp["EpisodeNo"] / george_grp["LineCount"]
george_grp["Season"] = george_grp["Season"] / george_grp["LineCount"]
george_grp["PosSent"] = george_grp["WtdPos"] / george_grp["WordCount"]
george_grp["NegSent"] = george_grp["WtdNeg"] / george_grp["WordCount"]
george_grp["CompSent"] = george_grp["WtdComp"] / george_grp["WordCount"]

# Drop column for WtdSentiments since we no longer need it
george_grp = george_grp.drop("WtdPos", 1)
george_grp = george_grp.drop("WtdNeg", 1)
george_grp = george_grp.drop("WtdComp", 1)

# Rename Sentiment, LineCount, and WordCount variables to have "Total" at the end of them
george_grp = george_grp.rename(index=str, columns = {"PosSent": "PosGeorge",
                                                       "NegSent": "NegGeorge",
                                                       "CompSent": "CompGeorge",
                                                       "LineCount": "LineCountGeorge",
                                                       "WordCount": "WordCountGeorge"
                                                      })
george_grp.tail()

# Grouping by Episode -- Kramer
kramer_grp = kramer_df.groupby(by = "SEID", as_index=False).sum()

# Drop the EpisodeNo, Season, and Sentiment columns since we don't need them anymore

# Re-create columns for EpisodeNo, Season and Sentiment
kramer_grp["EpisodeNo"] = kramer_grp["EpisodeNo"] / kramer_grp["LineCount"]
kramer_grp["Season"] = kramer_grp["Season"] / kramer_grp["LineCount"]
kramer_grp["PosSent"] = kramer_grp["WtdPos"] / kramer_grp["WordCount"]
kramer_grp["NegSent"] = kramer_grp["WtdNeg"] / kramer_grp["WordCount"]
kramer_grp["CompSent"] = kramer_grp["WtdComp"] / kramer_grp["WordCount"]

# Drop column for WtdSentiments since we no longer need it
kramer_grp = kramer_grp.drop("WtdPos", 1)
kramer_grp = kramer_grp.drop("WtdNeg", 1)
kramer_grp = kramer_grp.drop("WtdComp", 1)

# Rename Sentiment, LineCount, and WordCount variables to have "Total" at the end of them
kramer_grp = kramer_grp.rename(index=str, columns = {"PosSent": "PosKramer",
                                                       "NegSent": "NegKramer",
                                                       "CompSent": "CompKramer",
                                                       "LineCount": "LineCountKramer",
                                                       "WordCount": "WordCountKramer"
                                                      })
kramer_grp.tail()

# Grouping by Episode -- Elaine
elaine_grp = elaine_df.groupby(by = "SEID", as_index=False).sum()

# Drop the EpisodeNo, Season, and Sentiment columns since we don't need them anymore

# Re-create columns for EpisodeNo, Season and Sentiment
elaine_grp["EpisodeNo"] = elaine_grp["EpisodeNo"] / elaine_grp["LineCount"]
elaine_grp["Season"] = elaine_grp["Season"] / elaine_grp["LineCount"]
elaine_grp["PosSent"] = elaine_grp["WtdPos"] / elaine_grp["WordCount"]
elaine_grp["NegSent"] = elaine_grp["WtdNeg"] / elaine_grp["WordCount"]
elaine_grp["CompSent"] = elaine_grp["WtdComp"] / elaine_grp["WordCount"]

# Drop column for WtdSentiments since we no longer need it
elaine_grp = elaine_grp.drop("WtdPos", 1)
elaine_grp = elaine_grp.drop("WtdNeg", 1)
elaine_grp = elaine_grp.drop("WtdComp", 1)

# Rename Sentiment, LineCount, and WordCount variables to have "Total" at the end of them
elaine_grp = elaine_grp.rename(index=str, columns = {"PosSent": "PosElaine",
                                                       "NegSent": "NegElaine",
                                                       "CompSent": "CompElaine",
                                                       "LineCount": "LineCountElaine",
                                                       "WordCount": "WordCountElaine"
                                                      })
elaine_grp.tail()

# Grouping by Episode -- Secondary characters
secondary_grp = secondary_df.groupby(by = "SEID", as_index=False).sum()

# Drop the EpisodeNo, Season, and Sentiment columns since we don't need them anymore

# Re-create columns for EpisodeNo, Season and Sentiment
secondary_grp["EpisodeNo"] = secondary_grp["EpisodeNo"] / secondary_grp["LineCount"]
secondary_grp["Season"] = secondary_grp["Season"] / secondary_grp["LineCount"]
secondary_grp["PosSent"] = secondary_grp["WtdPos"] / secondary_grp["WordCount"]
secondary_grp["NegSent"] = secondary_grp["WtdNeg"] / secondary_grp["WordCount"]
secondary_grp["CompSent"] = secondary_grp["WtdComp"] / secondary_grp["WordCount"]

# Drop column for WtdSentiments since we no longer need it
secondary_grp = secondary_grp.drop("WtdPos", 1)
secondary_grp = secondary_grp.drop("WtdNeg", 1)
secondary_grp = secondary_grp.drop("WtdComp", 1)

# Rename Sentiment, LineCount, and WordCount variables to have "Total" at the end of them
secondary_grp = secondary_grp.rename(index=str, columns = {"PosSent": "PosSecond",
                                                       "NegSent": "NegSecond",
                                                       "CompSent": "CompSecond",
                                                       "LineCount": "LineCountSecond",
                                                       "WordCount": "WordCountSecond"
                                                      })
secondary_grp.tail()

# Grouping by Episode -- Other characters
other_grp = other_df.groupby(by = "SEID", as_index=False).sum()

# Drop the EpisodeNo, Season, and Sentiment columns since we don't need them anymore

# Re-create columns for EpisodeNo, Season and Sentiment
other_grp["EpisodeNo"] = other_grp["EpisodeNo"] / other_grp["LineCount"]
other_grp["Season"] = other_grp["Season"] / other_grp["LineCount"]
other_grp["PosSent"] = other_grp["WtdPos"] / other_grp["WordCount"]
other_grp["NegSent"] = other_grp["WtdNeg"] / other_grp["WordCount"]
other_grp["CompSent"] = other_grp["WtdComp"] / other_grp["WordCount"]

# Drop column for WtdSentiments since we no longer need it
other_grp = other_grp.drop("WtdPos", 1)
other_grp = other_grp.drop("WtdNeg", 1)
other_grp = other_grp.drop("WtdComp", 1)

# Rename Sentiment, LineCount, and WordCount variables to have "Total" at the end of them
other_grp = other_grp.rename(index=str, columns = {"PosSent": "PosOther",
                                                       "NegSent": "NegOther",
                                                       "CompSent": "CompOther",
                                                       "LineCount": "LineCountOther",
                                                       "WordCount": "WordCountOther"
                                                      })
other_grp.tail()

##### Merge all grouped dataframes into one #####
# overall_grp
# jerry_grp
# george_grp
# kramer_grp
# elaine_grp
# secondary_grp
# other_grp

# Merging Overall and Jerry
merged_df = pd.merge(overall_grp, jerry_grp, on = "SEID", how = "left")

# Drop EpisodeNo_y and Season_y
merged_df = merged_df.drop("EpisodeNo_y", 1)
merged_df = merged_df.drop("Season_y", 1)

# Rename EpsodeNo_x and Season_x
merged_df = merged_df.rename(index=str, columns = {"EpisodeNo_x": "EpisodeNo", "Season_x": "Season"})

#print(merged_df.shape)
merged_df.head()

# Merging above with George
merged_df = pd.merge(merged_df, george_grp, on = "SEID", how = "left")

# Drop EpisodeNo_y and Season_y
merged_df = merged_df.drop("EpisodeNo_y", 1)
merged_df = merged_df.drop("Season_y", 1)

# Rename EpsodeNo_x and Season_x
merged_df = merged_df.rename(index=str, columns = {"EpisodeNo_x": "EpisodeNo", "Season_x": "Season"})

print(merged_df.shape)
print(merged_df.head())

# Merging above with Kramer
merged_df = pd.merge(merged_df, kramer_grp, on = "SEID", how = "left")

# Drop EpisodeNo_y and Season_y
merged_df = merged_df.drop("EpisodeNo_y", 1)
merged_df = merged_df.drop("Season_y", 1)

# Rename EpsodeNo_x and Season_x
merged_df = merged_df.rename(index=str, columns = {"EpisodeNo_x": "EpisodeNo", "Season_x": "Season"})

print(merged_df.shape)
print(merged_df.head())

# Merging above with Elaine
merged_df = pd.merge(merged_df, elaine_grp, on = "SEID", how = "left")

# Drop EpisodeNo_y and Season_y
merged_df = merged_df.drop("EpisodeNo_y", 1)
merged_df = merged_df.drop("Season_y", 1)

# Rename EpsodeNo_x and Season_x
merged_df = merged_df.rename(index=str, columns = {"EpisodeNo_x": "EpisodeNo", "Season_x": "Season"})

print(merged_df.shape)
print(merged_df.head())

# Merging above with Secondary
merged_df = pd.merge(merged_df, secondary_grp, on = "SEID", how = "left")

# Drop EpisodeNo_y and Season_y
merged_df = merged_df.drop("EpisodeNo_y", 1)
merged_df = merged_df.drop("Season_y", 1)

# Rename EpsodeNo_x and Season_x
merged_df = merged_df.rename(index=str, columns = {"EpisodeNo_x": "EpisodeNo", "Season_x": "Season"})

print(merged_df.shape)
print(merged_df.head())

# Merging above with Other
merged_df = pd.merge(merged_df, other_grp, on = "SEID", how = "left")

# Drop EpisodeNo_y and Season_y
merged_df = merged_df.drop("EpisodeNo_y", 1)
merged_df = merged_df.drop("Season_y", 1)

# Rename EpsodeNo_x and Season_x
merged_df = merged_df.rename(index=str, columns = {"EpisodeNo_x": "EpisodeNo", "Season_x": "Season"})

print(merged_df.shape)
print(merged_df.head())

# Export to CSV
merged_df.to_csv('../outputData/merged_scripts.csv')
