==> python/3_final_merge.ipynb <==
# Import dependencies
import pandas as pd
import numpy as np

# Read in the CSV files
merged_df = pd.read_csv('../outputData/merged_scripts.csv')
episode_df = pd.read_csv('../rawData/episode_info.csv')

# Drop the "Unnamed: 0" and "Unnamed: 0.1" columns
merged_df = merged_df.drop("Unnamed: 0", 1)
episode_df = episode_df.drop("Unnamed: 0", 1)

# Drop S01E00, since we don't have any sentiment for that
episode_df = episode_df[1:episode_df.shape[0]]

merged_df.head()

episode_df.head()

### Merge both data frames on SEID
seinfeld_df = pd.merge(episode_df, merged_df, on = "SEID", how = "left")
seinfeld_df.head()

# Drop EpisodeNo_y and Season_y
seinfeld_df = seinfeld_df.drop("EpisodeNo_y", 1)
seinfeld_df = seinfeld_df.drop("Season_y", 1)

# Rename EpsodeNo_x and Season_x
seinfeld_df = seinfeld_df.rename(index=str, columns = {"EpisodeNo_x": "EpisodeNo", "Season_x": "Season"})
seinfeld_df.head()

# Rename "IMDB Rating" column to "IMDBrating"
seinfeld_df = seinfeld_df.rename(index=str, columns = {"IMDB Rating": "IMDBrating"})

#### Add dummy variables for two-part episodes
# S01E01
# S05E18
# S09E23

seinfeld_df["TwoParter"] = np.repeat(0, seinfeld_df.shape[0])

seinfeld_df.loc[seinfeld_df.SEID == "S01E01", "TwoParter"] = 1
seinfeld_df.loc[seinfeld_df.SEID == "S05E18", "TwoParter"] = 1
seinfeld_df.loc[seinfeld_df.SEID == "S09E23", "TwoParter"] = 1

seinfeld_df.head()

# Add separate dummy variables for the pilot and the finale
seinfeld_df["Pilot"] = np.repeat(0, seinfeld_df.shape[0])
seinfeld_df.loc[seinfeld_df.SEID == "S01E01", "Pilot"] = 1

seinfeld_df["Finale"] = np.repeat(0, seinfeld_df.shape[0])
seinfeld_df.loc[seinfeld_df.SEID == "S09E23", "Finale"] = 1

seinfeld_df.tail()

# Export to CSV
seinfeld_df.to_csv('../rawData/seinfeld.csv')
